Get rid of jq (#9)
Resolves #5 in another way.

We aren't going to install `jq` because it's too strange to install something from the script directly.
The output and `sudo` password request seems to be strange.
But from @mvovchuk words, an error message about not installed `jq` and instruction how to install it seems to be toooo difficult for the potential customers... That's why just don't use `jq` at all. `grep`, `tr` and `awk` are our friends.