Fix uninstall/install Elastio packages on RPM-based systems (#24)
There was an issue on CentOS / Amazon Linux when release packages are
installed and then uninstalled. And dev packages are installed with the
script param `-b master` but still release packages appeared in the
system. The issue was in the remaining files in the
/var/cache/yum/x86_64/2/Elastio/ after uninstall. The files repomd.xml
and mirrors.txt preserved link to the previously used repository.
Fixed this issue.