Extend instance profile validation (#32)
In case if someone deletes a vault, then he or she has an instance
profile with the role with the policy with access to the old removed
s3 bucket. But without access to the new bucket that he or she already
deleted.
I'm not going to add possibility to create a new policy and attach it to
the existing role, because this script is very temporary and will go
away soon. And here is a pretty easy workaround just to create a new
profile (and role and policy) to use new bucket.

Closes #2435