Install NBD on AL2 (#61)
AL2 has `nbd` kernel module but doesn't have `nbd-client`. It can be
installed from the `epel` repo as `nbd` package. But this repo is not
available for yum as on CentOS. That's why using AL2's way to enable
`epel` via their internal command and installing `nbd` package manually.
As result, the `nbd-client` version 3.14 is installed.

Resolves part 2 of https://github.com/elastio/elastio/issues/4295