ci: add tflint (#95)
This pull request introduces a new Terraform linting job to the CI
workflow and adds a configuration file for TFLint. The most important
changes include adding a new `tflint` job in the CI workflow and
configuring TFLint with specific plugins and settings.

### CI Workflow Enhancements:

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR88-R102):
Added a new `tflint` job that runs on `ubuntu-latest`, sets up TFLint,
and runs it with specific options.

### TFLint Configuration:

*
[`.tflint.hcl`](diffhunk://#diff-c82c2fd3fe5d1315facdf6794e11d51e0bf10783e6ef34a270996a62bccf605dR1-R25):
Added a new configuration file for TFLint specifying required version,
plugin directory, and enabling plugins for Terraform, AWS, and AzureRM
with specific versions and sources.

---------

Co-authored-by: Veetaha <gersoh3@gmail.com>