Release workflow (#93)
This pull request introduces a new script for releasing Terraform
modules and updates the GitHub Actions workflow to utilize this script.
The main changes include the addition of the `release-tf-module.sh`
script and updates to the `release.yml` workflow file.

### New script for releasing Terraform modules:

*
[`.github/scripts/release-tf-module.sh`](diffhunk://#diff-4bc2fa3625e5cb84c6cf8776ac345b8afe9c047f6730e22b685154a2afa9e615R1-R179):
Added a new script to package, publish, and tag Terraform modules. This
script includes functions for dependency checking, usage instructions,
and handling various options such as publishing, tagging, and cleanup.

### Updates to GitHub Actions workflow:

*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34R1-R43):
Updated the workflow to include a new job for releasing Terraform
modules. This job collects modules, configures Git, and runs the new
release script with appropriate options.